dev_requirements.txt
ruff
pytest-qgis